Continued with category grid design

+++ ExpenseCounter/View models/CategoryViewModel.swift
@@ -9,6 +9,9 @@ import CoreData
 
 class CategoryViewModel: ObservableObject {
     @Published var categories: [Category] = []
+    @Published var customCategories: [Category] = []
+    @Published var defaultCategories: [Category] = []
+    
     private let coreDataSharedInstance = CoreDataStack.shared
     
     init() {
@@ -31,14 +34,40 @@ class CategoryViewModel: ObservableObject {
         fetchCategories()
     }
     
+    private func searchCategory(_ id: UUID) -> Category? {
+        let fetchRequest: NSFetchRequest = Category.fetchRequest()
+        fetchRequest.predicate = NSPredicate(format: "id == %@", id as CVarArg)
+        fetchRequest.fetchLimit = 1
+        
+        do {
+            if let category = try coreDataSharedInstance.context.fetch(fetchRequest).first {
+                return category
+            }
+        } catch let error {
+            fatalError("Can't search for a category -> \(error.localizedDescription)")
+        }
+        return nil
+    }
+    
     func fetchCategories() {
         let request = NSFetchRequest<Category>(entityName: "Category")
         
         do {
             categories = try coreDataSharedInstance.context.fetch(request)
+            customCategories = categories.filter { $0.defaultCategory == false }
+            defaultCategories = categories.filter { $0.defaultCategory == true }
+            
         } catch let error {
             fatalError("Can't fetch categories with error -> \(error.localizedDescription)")
         }
+    }
+    
+    func getCustomCategories() {
+        customCategories = categories.filter { $0.defaultCategory == false }
+    }
+    
+    func getDefaultCategories() {
+        defaultCategories = categories.filter { $0.defaultCategory == true }
     }
     
     func addCategory(_ name: String, _ colorHex: String, _ icon: String) {
@@ -54,41 +83,29 @@ class CategoryViewModel: ObservableObject {
     }
     
     func updateCategory(_ id: UUID, _ name: String, _ colorHex: String, _ icon: String) {
-        let fetchRequest: NSFetchRequest = Category.fetchRequest()
-        fetchRequest.predicate = NSPredicate(format: "id == %@", id as CVarArg)
-        fetchRequest.fetchLimit = 1
+        let searchedCategory = searchCategory(id)
         
-        do {
-            if let existingCategory = try coreDataSharedInstance.context.fetch(fetchRequest).first {
-                existingCategory.name = name
-                existingCategory.colorHex = colorHex
-                existingCategory.icon = icon
-                
-                coreDataSharedInstance.save()
-                fetchCategories()
-            }
-        } catch let error {
-            fatalError("Error updating a category -> \(error.localizedDescription)")
+        if searchedCategory != nil {
+            searchedCategory?.name = name
+            searchedCategory?.colorHex = colorHex
+            searchedCategory?.icon = icon
+            
+            coreDataSharedInstance.save()
+            fetchCategories()
+        } else {
+            fatalError("Error updating a category")
         }
     }
     
-    func deleteCategory(_ category: Category) {
-        coreDataSharedInstance.context.delete(category)
-        coreDataSharedInstance.save()
-        fetchCategories()
+    func deleteCategory(_ id: UUID) {
+        let searchedCategory = searchCategory(id)
+        
+        if let category = searchedCategory {
+            coreDataSharedInstance.context.delete(category)
+            coreDataSharedInstance.save()
+            fetchCategories()
+        } else {
+            fatalError("Error deleting a category")
+        }
     }
-    
-    //Don't call this in production code
-//    private func deleteAllCategories() {
-//        let fetchRequest: NSFetchRequest<NSFetchRequestResult> = Category.fetchRequest()
-//        let batchDeleteRequest = NSBatchDeleteRequest(fetchRequest: fetchRequest)
-//
-//        do {
-//            try coreDataSharedInstance.context.execute(batchDeleteRequest)
-//            try coreDataSharedInstance.context.save()
-//            fetchCategories() // refresh the local categories array if needed
-//        } catch {
-//            print("Failed to delete all categories: \(error)")
-//        }
-//    }
 }
